Flytt springkomponenter som blir helt mocket i egne filer
