+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/cucumber/mock/CucumberMock.kt
@@ -1,56 +1,30 @@
 ﻿package no.nav.familie.ba.sak.cucumber.mock
 
-import io.mockk.every
-import io.mockk.just
 import io.mockk.mockk
-import io.mockk.runs
 import io.mockk.spyk
 import no.nav.familie.ba.sak.common.LocalDateProvider
-import no.nav.familie.ba.sak.common.lagVedtak
-import no.nav.familie.ba.sak.common.tilPersonEnkel
-import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
 import no.nav.familie.ba.sak.cucumber.BegrunnelseTeksterStepDefinition
 import no.nav.familie.ba.sak.integrasjoner.ecb.ECBService
 import no.nav.familie.ba.sak.integrasjoner.ef.EfSakRestClient
 import no.nav.familie.ba.sak.integrasjoner.infotrygd.InfotrygdService
-import no.nav.familie.ba.sak.integrasjoner.oppgave.OppgaveService
-import no.nav.familie.ba.sak.integrasjoner.pdl.PersonopplysningerService
-import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PersonInfo
 import no.nav.familie.ba.sak.integrasjoner.økonomi.UtbetalingsoppdragGenerator
 import no.nav.familie.ba.sak.integrasjoner.økonomi.UtbetalingsoppdragGeneratorService
-import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiKlient
 import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiService
-import no.nav.familie.ba.sak.kjerne.arbeidsfordeling.ArbeidsfordelingService
 import no.nav.familie.ba.sak.kjerne.autovedtak.AutovedtakService
 import no.nav.familie.ba.sak.kjerne.autovedtak.månedligvalutajustering.MånedligValutajusteringSevice
 import no.nav.familie.ba.sak.kjerne.autovedtak.småbarnstillegg.AutovedtakSmåbarnstilleggService
-import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
-import no.nav.familie.ba.sak.kjerne.behandling.BehandlingMetrikker
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
 import no.nav.familie.ba.sak.kjerne.behandling.SnikeIKøenService
 import no.nav.familie.ba.sak.kjerne.behandling.behandlingstema.BehandlingstemaService
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingMigreringsinfoRepository
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingStatus
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingSøknadsinfoService
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
-import no.nav.familie.ba.sak.kjerne.behandling.settpåvent.SettPåVentRepository
 import no.nav.familie.ba.sak.kjerne.behandling.settpåvent.SettPåVentService
 import no.nav.familie.ba.sak.kjerne.behandlingsresultat.BehandlingsresultatService
 import no.nav.familie.ba.sak.kjerne.behandlingsresultat.BehandlingsresultatSteg
 import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
 import no.nav.familie.ba.sak.kjerne.beregning.SmåbarnstilleggService
 import no.nav.familie.ba.sak.kjerne.beregning.TilkjentYtelseValideringService
-import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
-import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseRepository
-import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelerTilkjentYtelseOgEndreteUtbetalingerService
-import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelse
-import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelseRepository
-import no.nav.familie.ba.sak.kjerne.beregning.domene.tilAndelerTilkjentYtelseMedEndreteUtbetalinger
-import no.nav.familie.ba.sak.kjerne.beregning.domene.tilEndretUtbetalingAndelMedAndelerTilkjentYtelse
-import no.nav.familie.ba.sak.kjerne.brev.DokumentGenereringService
-import no.nav.familie.ba.sak.kjerne.endretutbetaling.EndretUtbetalingAndelHentOgPersisterService
-import no.nav.familie.ba.sak.kjerne.endretutbetaling.domene.EndretUtbetalingAndelRepository
 import no.nav.familie.ba.sak.kjerne.eøs.differanseberegning.TilpassDifferanseberegningEtterTilkjentYtelseService
 import no.nav.familie.ba.sak.kjerne.eøs.differanseberegning.TilpassDifferanseberegningEtterUtenlandskPeriodebeløpService
 import no.nav.familie.ba.sak.kjerne.eøs.differanseberegning.TilpassDifferanseberegningEtterValutakursService
@@ -58,81 +32,27 @@ import no.nav.familie.ba.sak.kjerne.eøs.differanseberegning.TilpassDifferansebe
 import no.nav.familie.ba.sak.kjerne.eøs.endringsabonnement.TilbakestillBehandlingFraKompetanseEndringService
 import no.nav.familie.ba.sak.kjerne.eøs.endringsabonnement.TilbakestillBehandlingFraUtenlandskPeriodebeløpEndringService
 import no.nav.familie.ba.sak.kjerne.eøs.endringsabonnement.TilbakestillBehandlingFraValutakursEndringService
-import no.nav.familie.ba.sak.kjerne.eøs.endringsabonnement.TilpassKompetanserTilRegelverkService
 import no.nav.familie.ba.sak.kjerne.eøs.endringsabonnement.TilpassUtenlandskePeriodebeløpTilKompetanserService
 import no.nav.familie.ba.sak.kjerne.eøs.endringsabonnement.TilpassValutakurserTilUtenlandskePeriodebeløpService
-import no.nav.familie.ba.sak.kjerne.eøs.felles.BehandlingId
-import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.KompetanseRepository
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.KompetanseService
-import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
-import no.nav.familie.ba.sak.kjerne.eøs.utenlandskperiodebeløp.UtenlandskPeriodebeløp
-import no.nav.familie.ba.sak.kjerne.eøs.utenlandskperiodebeløp.UtenlandskPeriodebeløpRepository
 import no.nav.familie.ba.sak.kjerne.eøs.utenlandskperiodebeløp.UtenlandskPeriodebeløpService
-import no.nav.familie.ba.sak.kjerne.eøs.valutakurs.Valutakurs
-import no.nav.familie.ba.sak.kjerne.eøs.valutakurs.ValutakursRepository
 import no.nav.familie.ba.sak.kjerne.eøs.valutakurs.ValutakursService
-import no.nav.familie.ba.sak.kjerne.eøs.vilkårsvurdering.VilkårsvurderingTidslinjeService
-import no.nav.familie.ba.sak.kjerne.fagsak.Fagsak
-import no.nav.familie.ba.sak.kjerne.fagsak.FagsakRepository
-import no.nav.familie.ba.sak.kjerne.fagsak.FagsakService
-import no.nav.familie.ba.sak.kjerne.fagsak.FagsakStatus
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlagRepository
-import no.nav.familie.ba.sak.kjerne.grunnlag.småbarnstillegg.PeriodeOvergangsstønadGrunnlag
-import no.nav.familie.ba.sak.kjerne.grunnlag.småbarnstillegg.PeriodeOvergangsstønadGrunnlagRepository
-import no.nav.familie.ba.sak.kjerne.grunnlag.søknad.SøknadGrunnlagService
-import no.nav.familie.ba.sak.kjerne.logg.Logg
-import no.nav.familie.ba.sak.kjerne.logg.LoggService
-import no.nav.familie.ba.sak.kjerne.logg.LoggType
-import no.nav.familie.ba.sak.kjerne.personident.Aktør
-import no.nav.familie.ba.sak.kjerne.personident.PersonidentService
-import no.nav.familie.ba.sak.kjerne.simulering.SimuleringService
-import no.nav.familie.ba.sak.kjerne.steg.BehandlerRolle
 import no.nav.familie.ba.sak.kjerne.steg.FerdigstillBehandling
 import no.nav.familie.ba.sak.kjerne.steg.IverksettMotOppdrag
 import no.nav.familie.ba.sak.kjerne.steg.RegistrerPersongrunnlag
 import no.nav.familie.ba.sak.kjerne.steg.StatusFraOppdrag
 import no.nav.familie.ba.sak.kjerne.steg.StegService
-import no.nav.familie.ba.sak.kjerne.steg.TilbakestillBehandlingService
 import no.nav.familie.ba.sak.kjerne.steg.TilbakestillBehandlingTilBehandlingsresultatService
 import no.nav.familie.ba.sak.kjerne.steg.VilkårsvurderingSteg
 import no.nav.familie.ba.sak.kjerne.steg.grunnlagForNyBehandling.EøsSkjemaerForNyBehandlingService
-import no.nav.familie.ba.sak.kjerne.steg.grunnlagForNyBehandling.PersonopplysningGrunnlagForNyBehandlingService
-import no.nav.familie.ba.sak.kjerne.steg.grunnlagForNyBehandling.VilkårsvurderingForNyBehandlingService
-import no.nav.familie.ba.sak.kjerne.tilbakekreving.TilbakekrevingService
-import no.nav.familie.ba.sak.kjerne.totrinnskontroll.TotrinnskontrollService
-import no.nav.familie.ba.sak.kjerne.totrinnskontroll.domene.Totrinnskontroll
-import no.nav.familie.ba.sak.kjerne.vedtak.Vedtak
-import no.nav.familie.ba.sak.kjerne.vedtak.VedtakRepository
 import no.nav.familie.ba.sak.kjerne.vedtak.VedtakService
-import no.nav.familie.ba.sak.kjerne.vedtak.domene.VedtaksperiodeMedBegrunnelser
-import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.VedtaksperiodeHentOgPersisterService
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.VedtaksperiodeService
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårService
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.VilkårsvurderingRepository
-import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
-import no.nav.familie.ba.sak.sikkerhet.TilgangService
-import no.nav.familie.ba.sak.statistikk.saksstatistikk.SaksstatistikkEventPublisher
 import no.nav.familie.ba.sak.task.FerdigstillBehandlingTask
 import no.nav.familie.ba.sak.task.IverksettMotOppdragTask
 import no.nav.familie.ba.sak.task.StatusFraOppdragTask
-import no.nav.familie.kontrakter.felles.ef.Datakilde
-import no.nav.familie.kontrakter.felles.ef.EksternePerioderResponse
-import no.nav.familie.kontrakter.felles.oppdrag.OppdragStatus
-import no.nav.familie.prosessering.AsyncTaskStep
-import no.nav.familie.prosessering.TaskStepBeskrivelse
-import no.nav.familie.prosessering.domene.Task
-import no.nav.familie.prosessering.internal.TaskService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.springframework.aop.framework.AopProxyUtils
-import org.springframework.core.annotation.AnnotationUtils
 import java.time.LocalDate
-import java.time.LocalDateTime
-import kotlin.random.Random
 
 val logger: Logger = LoggerFactory.getLogger("CucumberMock")
 
@@ -507,717 +427,6 @@ class CucumberMock(
     val taskservices = listOf(iverksettMotOppdragTask, ferdigstillBehandlingTask, statusFraOppdragTask)
 }
 
-private fun mockOppgaveService(): OppgaveService {
-    val oppgaveService = mockk<OppgaveService>()
-    every { oppgaveService.opprettOppgaveForManuellBehandling(any(), any(), any(), any()) } returns ""
-    return oppgaveService
-}
-
-fun mockSettPåVentRepository(): SettPåVentRepository {
-    val settPåVentRepository = mockk<SettPåVentRepository>()
-
-    every { settPåVentRepository.findByBehandlingIdAndAktiv(any(), any()) } returns null
-    return settPåVentRepository
-}
-
-private fun mockTilbakekrevingService(): TilbakekrevingService {
-    val tilbakekrevingService = mockk<TilbakekrevingService>()
-    every { tilbakekrevingService.slettTilbakekrevingPåBehandling(any()) } returns null
-    return tilbakekrevingService
-}
-
-private fun mockBehandlingMetrikker(): BehandlingMetrikker {
-    val behandlingMetrikker = mockk<BehandlingMetrikker>()
-    every { behandlingMetrikker.tellNøkkelTallVedOpprettelseAvBehandling(any()) } just runs
-    every { behandlingMetrikker.oppdaterBehandlingMetrikker(any()) } just runs
-    return behandlingMetrikker
-}
-
-private fun mockArbeidsfordelingService(): ArbeidsfordelingService {
-    val arbeidsfordelingService = mockk<ArbeidsfordelingService>()
-    every { arbeidsfordelingService.fastsettBehandlendeEnhet(any(), any()) } just runs
-    return arbeidsfordelingService
-}
-
-private fun mockSaksstatistikkEventPublisher(): SaksstatistikkEventPublisher {
-    val saksstatistikkEventPublisher = mockk<SaksstatistikkEventPublisher>()
-    every { saksstatistikkEventPublisher.publiserSaksstatistikk(any()) } just runs
-    every { saksstatistikkEventPublisher.publiserBehandlingsstatistikk(any()) } just runs
-    return saksstatistikkEventPublisher
-}
-
-private fun mockTaskService(): TaskService {
-    val taskService = mockk<TaskService>()
-    every { taskService.save(any()) } answers { firstArg() }
-    return taskService
-}
-
-private fun mockTilbakestillBehandlingService(): TilbakestillBehandlingService {
-    val tilbakestillBehandlingService = mockk<TilbakestillBehandlingService>()
-    every { tilbakestillBehandlingService.tilbakestillDataTilVilkårsvurderingssteg(any()) } just runs
-    return tilbakestillBehandlingService
-}
-
-private fun mockVilkårService(dataFraCucumber: BegrunnelseTeksterStepDefinition): VilkårService {
-    val vilkårService = mockk<VilkårService>()
-    every { vilkårService.hentVilkårsvurdering(any()) } answers {
-        val behandlingsId = firstArg<Long>()
-        dataFraCucumber.vilkårsvurderinger[behandlingsId]!!
-    }
-    every { vilkårService.hentVilkårsvurderingThrows(any()) } answers {
-        val behandlingsId = firstArg<Long>()
-        dataFraCucumber.vilkårsvurderinger[behandlingsId]!!
-    }
-    return vilkårService
-}
-
-private fun mockPersongrunnlagService(dataFraCucumber: BegrunnelseTeksterStepDefinition): PersongrunnlagService {
-    val persongrunnlagService = mockk<PersongrunnlagService>()
-    every { persongrunnlagService.hentSøkerOgBarnPåBehandlingThrows(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        val personopplysningGrunnlag = dataFraCucumber.persongrunnlag[behandlingId] ?: error("Fant ikke persongrunnlag for behandling $behandlingId")
-        personopplysningGrunnlag.personer.map { it.tilPersonEnkel() }
-    }
-    every { persongrunnlagService.hentAktivThrows(any()) } answers {
-        val behandlingsId = firstArg<Long>()
-        dataFraCucumber.persongrunnlag[behandlingsId]!!
-    }
-    every { persongrunnlagService.hentBarna(any<Long>()) } answers {
-        val behandlingsId = firstArg<Long>()
-        dataFraCucumber.persongrunnlag[behandlingsId]!!.barna
-    }
-    return persongrunnlagService
-}
-
-private fun mockEøsSkjemaerForNyBehandlingService(): EøsSkjemaerForNyBehandlingService {
-    val eøsSkjemaerForNyBehandlingService = mockk<EøsSkjemaerForNyBehandlingService>()
-    every { eøsSkjemaerForNyBehandlingService.kopierEøsSkjemaer(any(), any()) } just runs
-    return eøsSkjemaerForNyBehandlingService
-}
-
-private fun mockVilkårsvurderingForNyBehandlingService(dataFraCucumber: BegrunnelseTeksterStepDefinition): VilkårsvurderingForNyBehandlingService {
-    val vilkårsvurderingForNyBehandlingService = mockk<VilkårsvurderingForNyBehandlingService>()
-    every { vilkårsvurderingForNyBehandlingService.opprettVilkårsvurderingUtenomHovedflyt(any(), any(), any()) } answers {
-        val behandling = firstArg<Behandling>()
-        val forrigeBehandling = secondArg<Behandling>()
-
-        val forrigeVilkårsvurdering = dataFraCucumber.vilkårsvurderinger[forrigeBehandling.id]!!
-        val vilkårsvurderingSmåbarnstilleggbehandling = forrigeVilkårsvurdering.copy(behandling = behandling).kopier()
-
-        dataFraCucumber.vilkårsvurderinger[behandling.id] = vilkårsvurderingSmåbarnstilleggbehandling
-    }
-    return vilkårsvurderingForNyBehandlingService
-}
-
-private fun mockTilgangService(): TilgangService {
-    val tilgangService = mockk<TilgangService>()
-    every { tilgangService.validerTilgangTilBehandling(any(), any()) } just runs
-    every { tilgangService.verifiserHarTilgangTilHandling(any(), any()) } just runs
-    return tilgangService
-}
-
-private fun mockPersonopplysningerService(dataFraCucumber: BegrunnelseTeksterStepDefinition): PersonopplysningerService {
-    val personopplysningerService = mockk<PersonopplysningerService>()
-    every { personopplysningerService.hentPersoninfoEnkel(any()) } answers {
-        val aktør = firstArg<Aktør>()
-        dataFraCucumber.persongrunnlag.values.flatMap { it.personer }.first { it.aktør == aktør }.tilPersonInfo()
-    }
-    return personopplysningerService
-}
-
-private fun mockFagsakService(dataFraCucumber: BegrunnelseTeksterStepDefinition): FagsakService {
-    val fagsakService = mockk<FagsakService>()
-    every { fagsakService.hentNormalFagsak(any()) } answers {
-        val aktør = firstArg<Aktør>()
-        dataFraCucumber.fagsaker.values.single { it.aktør == aktør }
-    }
-    every { fagsakService.hentFagsakerPåPerson(any()) } answers {
-        listOf(dataFraCucumber.fagsaker.values.first())
-    }
-    every { fagsakService.oppdaterStatus(any(), any()) } answers {
-        val fagsak = firstArg<Fagsak>()
-        val nyStatus = secondArg<FagsakStatus>()
-
-        fagsak.status = nyStatus
-        fagsak
-    }
-    return fagsakService
-}
-
-private fun mockSnikeIKøenService(): SnikeIKøenService {
-    val snikeIKøenService = mockk<SnikeIKøenService>()
-    every { snikeIKøenService.kanSnikeForbi(any()) } answers { true }
-    return snikeIKøenService
-}
-
-private fun mockFagsakRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): FagsakRepository {
-    val fagsakRepository = mockk<FagsakRepository>()
-    every { fagsakRepository.finnFagsak(any()) } answers {
-        val id = firstArg<Long>()
-        dataFraCucumber.fagsaker.values.single { it.id == id }
-    }
-    return fagsakRepository
-}
-
-private fun mockTotrinnskontrollService(dataFraCucumber: BegrunnelseTeksterStepDefinition): TotrinnskontrollService {
-    val totrinnskontrollService = mockk<TotrinnskontrollService>()
-    every { totrinnskontrollService.opprettAutomatiskTotrinnskontroll(any()) } just runs
-    every { totrinnskontrollService.hentAktivForBehandling(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        Totrinnskontroll(
-            behandling = dataFraCucumber.behandlinger[behandlingId]!!,
-            godkjent = true,
-            saksbehandler = SikkerhetContext.SYSTEM_NAVN,
-            saksbehandlerId = SikkerhetContext.SYSTEM_FORKORTELSE,
-            beslutter = SikkerhetContext.SYSTEM_NAVN,
-            beslutterId = SikkerhetContext.SYSTEM_FORKORTELSE,
-        )
-    }
-    return totrinnskontrollService
-}
-
-private fun mockSimuleringService(): SimuleringService {
-    val simuleringService = mockk<SimuleringService>()
-    every { simuleringService.oppdaterSimuleringPåBehandling(any()) } returns emptyList()
-    return simuleringService
-}
-
-private fun mockEndretUtbetalingAndelRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): EndretUtbetalingAndelRepository {
-    val endretUtbetalingAndelRepository = mockk<EndretUtbetalingAndelRepository>()
-    every { endretUtbetalingAndelRepository.findByBehandlingId(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.endredeUtbetalinger[behandlingId] ?: emptyList()
-    }
-    return endretUtbetalingAndelRepository
-}
-
-private fun mockUtenlandskPeriodebeløpRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): UtenlandskPeriodebeløpRepository {
-    val utenlandskPeriodebeløpRepository = mockk<UtenlandskPeriodebeløpRepository>()
-    every { utenlandskPeriodebeløpRepository.finnFraBehandlingId(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.utenlandskPeriodebeløp[behandlingId] ?: emptyList()
-    }
-    every { utenlandskPeriodebeløpRepository.deleteAll(any<Iterable<UtenlandskPeriodebeløp>>()) } answers {
-        val utenlandskPeriodebeløp = firstArg<Iterable<UtenlandskPeriodebeløp>>()
-        utenlandskPeriodebeløp.forEach {
-            dataFraCucumber.utenlandskPeriodebeløp[it.behandlingId] = dataFraCucumber.utenlandskPeriodebeløp[it.behandlingId]?.filter { utenlandskPeriodebeløp -> utenlandskPeriodebeløp != it } ?: emptyList()
-        }
-    }
-    every { utenlandskPeriodebeløpRepository.saveAll(any<Iterable<UtenlandskPeriodebeløp>>()) } answers {
-        val utenlandskPeriodebeløp = firstArg<Iterable<UtenlandskPeriodebeløp>>()
-        utenlandskPeriodebeløp.forEach {
-            dataFraCucumber.utenlandskPeriodebeløp[it.behandlingId] = (dataFraCucumber.utenlandskPeriodebeløp[it.behandlingId] ?: emptyList()) + it
-        }
-        utenlandskPeriodebeløp.toList()
-    }
-    return utenlandskPeriodebeløpRepository
-}
-
-private fun mockValutakursRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): ValutakursRepository {
-    val valutakursRepository = mockk<ValutakursRepository>()
-    every { valutakursRepository.finnFraBehandlingId(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.valutakurs[behandlingId] ?: emptyList()
-    }
-    every { valutakursRepository.deleteAll(any<Iterable<Valutakurs>>()) } answers {
-        val valutakurser = firstArg<Iterable<Valutakurs>>()
-        valutakurser.forEach {
-            dataFraCucumber.valutakurs[it.behandlingId] = dataFraCucumber.valutakurs[it.behandlingId]?.filter { valutakurs -> valutakurs != it } ?: emptyList()
-        }
-    }
-    every { valutakursRepository.saveAll(any<Iterable<Valutakurs>>()) } answers {
-        val valutakurser = firstArg<Iterable<Valutakurs>>()
-        valutakurser.forEach {
-            dataFraCucumber.valutakurs[it.behandlingId] = (dataFraCucumber.valutakurs[it.behandlingId] ?: emptyList()) + it
-        }
-        valutakurser.toList()
-    }
-
-    return valutakursRepository
-}
-
-private fun mockKompetanseRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): KompetanseRepository {
-    val kompetanseRepository = mockk<KompetanseRepository>()
-    every { kompetanseRepository.finnFraBehandlingId(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.kompetanser[behandlingId] ?: emptyList()
-    }
-    every { kompetanseRepository.deleteAll(any<Iterable<Kompetanse>>()) } answers {
-        val kompetanser = firstArg<Iterable<Kompetanse>>()
-        kompetanser.forEach {
-            dataFraCucumber.kompetanser[it.behandlingId] = dataFraCucumber.kompetanser[it.behandlingId]?.filter { kompetanse -> kompetanse != it } ?: emptyList()
-        }
-    }
-    every { kompetanseRepository.saveAll(any<Iterable<Kompetanse>>()) } answers {
-        val kompetanser = firstArg<Iterable<Kompetanse>>()
-        kompetanser.forEach {
-            dataFraCucumber.kompetanser[it.behandlingId] = (dataFraCucumber.kompetanser[it.behandlingId] ?: emptyList()) + it
-        }
-        kompetanser.toMutableList()
-    }
-    return kompetanseRepository
-}
-
-private fun mockVedtaksperiodeHentOgPersisterService(dataFraCucumber: BegrunnelseTeksterStepDefinition): VedtaksperiodeHentOgPersisterService {
-    val vedtaksperiodeHentOgPersisterService = mockk<VedtaksperiodeHentOgPersisterService>()
-    every { vedtaksperiodeHentOgPersisterService.slettVedtaksperioderFor(any()) } answers {
-        val vedtak = firstArg<Vedtak>()
-        dataFraCucumber.vedtaksperioderMedBegrunnelser = dataFraCucumber.vedtaksperioderMedBegrunnelser.filter { it.vedtak.id != vedtak.id }
-    }
-    every { vedtaksperiodeHentOgPersisterService.lagre(any<List<VedtaksperiodeMedBegrunnelser>>()) } answers {
-        val vedtaksperioder = firstArg<List<VedtaksperiodeMedBegrunnelser>>()
-        dataFraCucumber.vedtaksperioderMedBegrunnelser = dataFraCucumber.vedtaksperioderMedBegrunnelser + vedtaksperioder
-        vedtaksperioder
-    }
-    every { vedtaksperiodeHentOgPersisterService.lagre(any<VedtaksperiodeMedBegrunnelser>()) } answers {
-        val vedtaksperiode = firstArg<VedtaksperiodeMedBegrunnelser>()
-        dataFraCucumber.vedtaksperioderMedBegrunnelser = dataFraCucumber.vedtaksperioderMedBegrunnelser + vedtaksperiode
-        vedtaksperiode
-    }
-    every { vedtaksperiodeHentOgPersisterService.finnVedtaksperioderFor(any()) } answers {
-        val vedtakId = firstArg<Long>()
-        dataFraCucumber.vedtaksperioderMedBegrunnelser.filter { it.vedtak.id == vedtakId }
-    }
-    return vedtaksperiodeHentOgPersisterService
-}
-
-private fun mockDokumentGenereringService(): DokumentGenereringService {
-    val dokumentGenereringService = mockk<DokumentGenereringService>()
-    every { dokumentGenereringService.genererBrevForVedtak(any()) } returns byteArrayOf()
-    return dokumentGenereringService
-}
-
-private fun mockVedtakRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): VedtakRepository {
-    val vedtakRepository = mockk<VedtakRepository>()
-    every { vedtakRepository.findByBehandlingAndAktiv(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        opprettEllerHentVedtak(dataFraCucumber, behandlingId)
-    }
-    every { vedtakRepository.getReferenceById(any()) } answers {
-        val vedtakId = firstArg<Long>()
-        dataFraCucumber.vedtaksliste.first { it.id == vedtakId }
-    }
-    every { vedtakRepository.findByBehandlingAndAktivOptional(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        opprettEllerHentVedtak(dataFraCucumber, behandlingId)
-    }
-    every { vedtakRepository.save(any()) } answers {
-        val oppdatertVedtak = firstArg<Vedtak>()
-        lagreVedtak(dataFraCucumber, oppdatertVedtak)
-    }
-    every { vedtakRepository.saveAndFlush(any()) } answers {
-        val oppdatertVedtak = firstArg<Vedtak>()
-        lagreVedtak(dataFraCucumber, oppdatertVedtak)
-    }
-    return vedtakRepository
-}
-
-private fun lagreVedtak(
-    dataFraCucumber: BegrunnelseTeksterStepDefinition,
-    oppdatertVedtak: Vedtak,
-): Vedtak {
-    dataFraCucumber.vedtaksliste = dataFraCucumber.vedtaksliste.map { if (it.id == oppdatertVedtak.id) oppdatertVedtak else it }.toMutableList()
-    if (oppdatertVedtak.id !in dataFraCucumber.vedtaksliste.map { it.id }) {
-        dataFraCucumber.vedtaksliste.add(oppdatertVedtak)
-    }
-    return oppdatertVedtak
-}
-
-private fun mockKompetanseService(dataFraCucumber: BegrunnelseTeksterStepDefinition): KompetanseService {
-    val kompetanseService = mockk<KompetanseService>()
-    every { kompetanseService.hentKompetanser(any()) } answers {
-        val behandlingId = firstArg<BehandlingId>()
-        dataFraCucumber.kompetanser[behandlingId.id] ?: emptyList()
-    }
-    return kompetanseService
-}
-
-private fun mockEndretUtbetalingAndelHentOgPersisterService(dataFraCucumber: BegrunnelseTeksterStepDefinition): EndretUtbetalingAndelHentOgPersisterService {
-    val endretUtbetalingAndelHentOgPersisterService = mockk<EndretUtbetalingAndelHentOgPersisterService>()
-    every { endretUtbetalingAndelHentOgPersisterService.hentForBehandling(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.endredeUtbetalinger[behandlingId] ?: emptyList()
-    }
-    return endretUtbetalingAndelHentOgPersisterService
-}
-
-private fun mockSøknadGrunnlagService(dataFraCucumber: BegrunnelseTeksterStepDefinition): SøknadGrunnlagService {
-    val søknadGrunnlagService = mockk<SøknadGrunnlagService>()
-    every { søknadGrunnlagService.hentAktiv(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        val behandling = dataFraCucumber.behandlinger[behandlingId]!!
-        if (behandling.opprettetÅrsak == BehandlingÅrsak.SØKNAD) {
-            error("Ikke implementert")
-        } else {
-            null
-        }
-    }
-    return søknadGrunnlagService
-}
-
-private fun mockTilpassKompetanserTilRegelverkService(): TilpassKompetanserTilRegelverkService {
-    val tilpassKompetanserTilRegelverkService = mockk<TilpassKompetanserTilRegelverkService>()
-    every { tilpassKompetanserTilRegelverkService.tilpassKompetanserTilRegelverk(any()) } just runs
-    return tilpassKompetanserTilRegelverkService
-}
-
-private fun mockPeriodeOvergangsstønadGrunnlagRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): PeriodeOvergangsstønadGrunnlagRepository {
-    val periodeOvergangsstønadGrunnlagRepository = mockk<PeriodeOvergangsstønadGrunnlagRepository>()
-    every { periodeOvergangsstønadGrunnlagRepository.deleteByBehandlingId(any()) } just runs
-    every { periodeOvergangsstønadGrunnlagRepository.saveAll(any<List<PeriodeOvergangsstønadGrunnlag>>()) } answers {
-        val overgangstønadsperioder = firstArg<List<PeriodeOvergangsstønadGrunnlag>>()
-        val behandlingId = overgangstønadsperioder.firstOrNull()?.behandlingId
-
-        if (behandlingId != null) {
-            dataFraCucumber.overgangsstønader[behandlingId] = overgangstønadsperioder.map { it.tilInternPeriodeOvergangsstønad() }
-        }
-
-        overgangstønadsperioder
-    }
-    every { periodeOvergangsstønadGrunnlagRepository.findByBehandlingId(any()) } answers {
-        val behandlingId = firstArg<Long>()
-
-        dataFraCucumber.overgangsstønader[behandlingId]?.map {
-            PeriodeOvergangsstønadGrunnlag(
-                behandlingId = behandlingId,
-                aktør = dataFraCucumber.persongrunnlag[behandlingId]!!.søker.aktør,
-                fom = it.fomDato,
-                tom = it.tomDato,
-                datakilde = Datakilde.EF,
-            )
-        } ?: emptyList()
-    }
-    return periodeOvergangsstønadGrunnlagRepository
-}
-
-private fun mockLoggService(): LoggService {
-    val loggService = mockk<LoggService>()
-    every { loggService.opprettBeslutningOmVedtakLogg(any(), any(), any(), any()) } just runs
-    every { loggService.opprettBehandlingLogg(any()) } just runs
-    every { loggService.opprettSettPåMaskinellVent(any(), any()) } just runs
-    every { loggService.opprettVilkårsvurderingLogg(any(), any(), any()) } returns null
-    every { loggService.opprettFerdigstillBehandling(any()) } just runs
-    every { loggService.opprettTattAvMaskinellVent(any()) } just runs
-    every { loggService.hentLoggForBehandling(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        listOf(
-            Logg(
-                behandlingId = behandlingId,
-                type = LoggType.BEHANDLING_OPPRETTET,
-                rolle = BehandlerRolle.SYSTEM,
-                tekst = "",
-            ).copy(opprettetTidspunkt = LocalDateTime.now().minusDays(1)),
-        )
-    }
-    return loggService
-}
-
-private fun mockVilkårsvurderingTidslinjeService(
-    vilkårsvurderingRepository: VilkårsvurderingRepository,
-    vilkårsvurderingService: VilkårsvurderingService,
-    persongrunnlagService: PersongrunnlagService,
-): VilkårsvurderingTidslinjeService {
-    val vilkårsvurderingTidslinjeService =
-        VilkårsvurderingTidslinjeService(
-            vilkårsvurderingRepository = vilkårsvurderingRepository,
-            vilkårsvurderingService = vilkårsvurderingService,
-            persongrunnlagService = persongrunnlagService,
-        )
-    return vilkårsvurderingTidslinjeService
-}
-
-private fun mockVilkårsvurderingService(dataFraCucumber: BegrunnelseTeksterStepDefinition): VilkårsvurderingService {
-    val vilkårsvurderingService = mockk<VilkårsvurderingService>()
-    every { vilkårsvurderingService.hentAktivForBehandling(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.vilkårsvurderinger[behandlingId]!!
-    }
-    every { vilkårsvurderingService.hentAktivForBehandlingThrows(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.vilkårsvurderinger[behandlingId]!!
-    }
-    return vilkårsvurderingService
-}
-
-private fun mockAndelTilkjentYtelseRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): AndelTilkjentYtelseRepository {
-    val andelTilkjentYtelseRepository = mockk<AndelTilkjentYtelseRepository>()
-    every { andelTilkjentYtelseRepository.finnAndelerTilkjentYtelseForBehandling(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.tilkjenteYtelser[behandlingId]?.andelerTilkjentYtelse?.toList() ?: emptyList()
-    }
-    every { andelTilkjentYtelseRepository.hentSisteAndelPerIdentOgType(any()) } answers {
-        val fagsakId = firstArg<Long>()
-        val behandlingId = dataFraCucumber.behandlinger.filter { it.value.fagsak.id == fagsakId }.filter { it.value.status == BehandlingStatus.AVSLUTTET }.maxByOrNull { it.value.id }?.key
-        val andelerPåBehandling = dataFraCucumber.tilkjenteYtelser[behandlingId]?.andelerTilkjentYtelse ?: emptyList()
-        andelerPåBehandling.tilSisteAndelPerAktørOgType()
-    }
-    return andelTilkjentYtelseRepository
-}
-
-fun Collection<AndelTilkjentYtelse>.tilSisteAndelPerAktørOgType(): List<AndelTilkjentYtelse> {
-    val andelerGruppertPåAktørOgType = groupBy { Pair(it.aktør, it.type) }
-    val sisteAndelPerAktørOgType = andelerGruppertPåAktørOgType.map { it.value.maxBy { it.stønadFom } }
-    return sisteAndelPerAktørOgType
-}
-
-private fun mockAndelerTilkjentYtelseOgEndreteUtbetalingerService(dataFraCucumber: BegrunnelseTeksterStepDefinition): AndelerTilkjentYtelseOgEndreteUtbetalingerService {
-    val andelerTilkjentYtelseOgEndreteUtbetalingerService = mockk<AndelerTilkjentYtelseOgEndreteUtbetalingerService>()
-    every { andelerTilkjentYtelseOgEndreteUtbetalingerService.finnAndelerTilkjentYtelseMedEndreteUtbetalinger(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        val andelerTilkjentYtelse = dataFraCucumber.tilkjenteYtelser[behandlingId]?.andelerTilkjentYtelse?.toList() ?: emptyList()
-        val endredeUtbetalinger = dataFraCucumber.endredeUtbetalinger[behandlingId] ?: emptyList()
-
-        andelerTilkjentYtelse.tilAndelerTilkjentYtelseMedEndreteUtbetalinger(endredeUtbetalinger)
-    }
-    every {
-        andelerTilkjentYtelseOgEndreteUtbetalingerService.finnEndreteUtbetalingerMedAndelerTilkjentYtelse(any())
-    } answers {
-        val behandlingId = firstArg<Long>()
-        val andelerTilkjentYtelse = dataFraCucumber.tilkjenteYtelser[behandlingId]?.andelerTilkjentYtelse?.toList() ?: emptyList()
-        val endredeUtbetalinger = dataFraCucumber.endredeUtbetalinger[behandlingId] ?: emptyList()
-
-        endredeUtbetalinger.tilEndretUtbetalingAndelMedAndelerTilkjentYtelse(andelerTilkjentYtelse)
-    }
-    return andelerTilkjentYtelseOgEndreteUtbetalingerService
-}
-
-private fun mockVilkårsvurderingRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): VilkårsvurderingRepository {
-    val vilkårsvurderingRepository = mockk<VilkårsvurderingRepository>()
-    every { vilkårsvurderingRepository.findByBehandlingAndAktiv(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.vilkårsvurderinger[behandlingId]!!
-    }
-    return vilkårsvurderingRepository
-}
-
-private fun mockEfSakRestClient(): EfSakRestClient {
-    val efSakRestClient = mockk<EfSakRestClient>()
-    every { efSakRestClient.hentPerioderMedFullOvergangsstønad(any()) } answers {
-        EksternePerioderResponse(emptyList())
-    }
-    return efSakRestClient
-}
-
-private fun mockTilkjentYtelseRepository(dataFraCucumber: BegrunnelseTeksterStepDefinition): TilkjentYtelseRepository {
-    val tilkjentYtelseRepository = mockk<TilkjentYtelseRepository>()
-    every { tilkjentYtelseRepository.findByBehandling(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.tilkjenteYtelser[behandlingId]!!
-    }
-    every { tilkjentYtelseRepository.findByBehandlingOptional(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.tilkjenteYtelser[behandlingId]
-    }
-    every { tilkjentYtelseRepository.slettTilkjentYtelseFor(any()) } just runs
-    every { tilkjentYtelseRepository.save(any()) } answers {
-        val tilkjentYtelse = firstArg<TilkjentYtelse>()
-        dataFraCucumber.tilkjenteYtelser[tilkjentYtelse.behandling.id] = tilkjentYtelse
-        tilkjentYtelse
-    }
-    every { tilkjentYtelseRepository.saveAndFlush(any()) } answers {
-        val tilkjentYtelse = firstArg<TilkjentYtelse>()
-
-        val tilkjentYtelseMedUnikIdPåAndeler = tilkjentYtelse.copy(andelerTilkjentYtelse = tilkjentYtelse.andelerTilkjentYtelse.map { it.copy(id = Random.nextLong()) }.toMutableSet())
-        dataFraCucumber.tilkjenteYtelser[tilkjentYtelse.behandling.id] = tilkjentYtelseMedUnikIdPåAndeler
-        tilkjentYtelseMedUnikIdPåAndeler
-    }
-    every { tilkjentYtelseRepository.findByBehandlingAndHasUtbetalingsoppdrag(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.tilkjenteYtelser[behandlingId]
-    }
-    return tilkjentYtelseRepository
-}
-
-private fun mockPersonidentService(
-    dataFraCucumber: BegrunnelseTeksterStepDefinition,
-    småbarnstilleggBehandlingId: Long,
-): PersonidentService {
-    val personidentService = mockk<PersonidentService>()
-    every { personidentService.hentOgLagreAktør(any(), any()) } answers {
-        val personIdent = firstArg<String>()
-        dataFraCucumber.persongrunnlag[småbarnstilleggBehandlingId]!!.personer.single { it.aktør.aktivFødselsnummer() == personIdent }.aktør
-    }
-    every { personidentService.hentAktør(any()) } answers {
-        val personIdent = firstArg<String>()
-        dataFraCucumber.persongrunnlag[småbarnstilleggBehandlingId]!!.personer.single { it.aktør.aktivFødselsnummer() == personIdent }.aktør
-    }
-    return personidentService
-}
-
-private fun mockPersonopplysningGrunnlagForNyBehandlingService(dataFraCucumber: BegrunnelseTeksterStepDefinition): PersonopplysningGrunnlagForNyBehandlingService {
-    val personopplysningGrunnlagForNyBehandlingService = mockk<PersonopplysningGrunnlagForNyBehandlingService>()
-    every { personopplysningGrunnlagForNyBehandlingService.opprettKopiEllerNyttPersonopplysningGrunnlag(any(), any(), any(), any()) } answers {
-        val behandling = firstArg<Behandling>()
-        val forrigeBehandling = secondArg<Behandling>()
-
-        val persongrunnlagForrigeBehandling = dataFraCucumber.persongrunnlag[forrigeBehandling.id]!!
-        dataFraCucumber.persongrunnlag[behandling.id] = persongrunnlagForrigeBehandling.copy(behandlingId = behandling.id)
-    }
-    return personopplysningGrunnlagForNyBehandlingService
-}
-
-private fun mockBehandlingHentOgPersisterService(
-    forrigeBehandling: Behandling?,
-    dataFraCucumber: BegrunnelseTeksterStepDefinition,
-    idForNyBehandling: Long,
-): BehandlingHentOgPersisterService {
-    val behandlingHentOgPersisterService = mockk<BehandlingHentOgPersisterService>()
-    every { behandlingHentOgPersisterService.hentSisteBehandlingSomErVedtatt(any()) } returns forrigeBehandling
-    every { behandlingHentOgPersisterService.hentForrigeBehandlingSomErIverksatt(any()) } returns forrigeBehandling
-    every { behandlingHentOgPersisterService.finnAktivOgÅpenForFagsak(any()) } answers {
-        val fagsakId = firstArg<Long>()
-        dataFraCucumber.behandlinger.values.singleOrNull { it.fagsak.id == fagsakId && it.status != BehandlingStatus.AVSLUTTET }
-    }
-    every { behandlingHentOgPersisterService.hentSisteBehandlingSomErIverksatt(any()) } answers {
-        val fagsakId = firstArg<Long>()
-        dataFraCucumber.behandlinger.values.filter { it.fagsak.id == fagsakId && it.status == BehandlingStatus.AVSLUTTET }.maxByOrNull { it.aktivertTidspunkt }
-    }
-    every { behandlingHentOgPersisterService.hentForrigeBehandlingSomErVedtatt(any()) } answers {
-        val behandling = firstArg<Behandling>()
-        dataFraCucumber.behandlinger.values.filter { it.fagsak.id == behandling.fagsak.id && it.id != behandling.id && it.status == BehandlingStatus.AVSLUTTET }.maxByOrNull { it.aktivertTidspunkt }
-    }
-    every { behandlingHentOgPersisterService.hent(any()) } answers {
-        val behandlingId = firstArg<Long>()
-        dataFraCucumber.behandlinger[behandlingId]!!
-    }
-    every { behandlingHentOgPersisterService.finnAktivForFagsak(any()) } answers {
-        forrigeBehandling
-    }
-    every { behandlingHentOgPersisterService.lagreOgFlush(any()) } answers {
-        val behandling = firstArg<Behandling>()
-        oppdaterEllerLagreBehandling(dataFraCucumber, behandling, idForNyBehandling)
-    }
-    every { behandlingHentOgPersisterService.lagreEllerOppdater(any(), any()) } answers {
-        val behandling = firstArg<Behandling>()
-        oppdaterEllerLagreBehandling(dataFraCucumber, behandling, idForNyBehandling)
-    }
-    every { behandlingHentOgPersisterService.hentBehandlinger(any()) } answers {
-        val fagsakId = firstArg<Long>()
-        dataFraCucumber.behandlinger.values.filter { it.fagsak.id == fagsakId }
-    }
-    every { behandlingHentOgPersisterService.hentBehandlinger(any(), any()) } answers {
-        val fagsakId = firstArg<Long>()
-        val status = secondArg<BehandlingStatus>()
-        dataFraCucumber.behandlinger.values.filter { it.fagsak.id == fagsakId && it.status == status }
-    }
-    return behandlingHentOgPersisterService
-}
-
-private fun oppdaterEllerLagreBehandling(
-    dataFraCucumber: BegrunnelseTeksterStepDefinition,
-    behandlingSomSkalLagres: Behandling,
-    idForNyBehandling: Long,
-): Behandling {
-    val behandling = if (behandlingSomSkalLagres.id == 0L) behandlingSomSkalLagres.copy(id = idForNyBehandling) else behandlingSomSkalLagres
-
-    dataFraCucumber.behandlinger[behandling.id] = behandling
-    return behandling
-}
-
-private fun mockVedtaksperiodeService(): VedtaksperiodeService {
-    val vedtaksperiodeService = mockk<VedtaksperiodeService>()
-    every { vedtaksperiodeService.oppdaterVedtakMedVedtaksperioder(any()) } just runs
-    return vedtaksperiodeService
-}
-
-private fun mockPersonopplysningGrunnlagRepository(behandlingIdTilPersongrunnlag: Map<Long, PersonopplysningGrunnlag>): PersonopplysningGrunnlagRepository {
-    val personopplysningGrunnlagRepository = mockk<PersonopplysningGrunnlagRepository>()
-    every { personopplysningGrunnlagRepository.findByBehandlingAndAktiv(any()) } answers {
-        val behandlingsId = firstArg<Long>()
-        behandlingIdTilPersongrunnlag[behandlingsId] ?: error("Fant ikke personopplysninggrunnlag for behandling $behandlingsId")
-    }
-    return personopplysningGrunnlagRepository
-}
-
-private fun mockØkonomiKlient(): ØkonomiKlient {
-    val økonomiKlient = mockk<ØkonomiKlient>()
-    every { økonomiKlient.iverksettOppdrag(any()) } returns ""
-    every { økonomiKlient.hentStatus(any()) } returns OppdragStatus.KVITTERT_OK
-    return økonomiKlient
-}
-
-private class MockTasker() {
-    val tasker = mutableListOf<Pair<AsyncTaskStep, Task>>()
-
-    // Sørger for at taskene blir kjørt i riktig rekkefølge og at vi ikke starter på en ny task før den forrige er ferdig.
-    fun kjørTaskOmIkkeAnnenTaskKjører(
-        taskStep: AsyncTaskStep,
-        task: Task,
-    ) {
-        tasker.add(taskStep to task)
-        if (tasker.size == 1) {
-            kjørNesteTask()
-        }
-    }
-
-    fun kjørNesteTask() {
-        if (tasker.isNotEmpty()) {
-            val (taskStep, task) = tasker.first()
-            taskStep.doTask(task)
-            taskStep.onCompletion(task)
-            tasker.removeAt(0)
-            if (tasker.isNotEmpty()) {
-                kjørNesteTask()
-            }
-        }
-    }
-
-    fun mockTaskRepositoryWrapper(cucumberMock: CucumberMock): TaskRepositoryWrapper {
-        val taskRepositoryWrapper = mockk<TaskRepositoryWrapper>()
-        every { taskRepositoryWrapper.save(any()) } answers {
-            val task = firstArg<Task>()
-            val tasktyper =
-                cucumberMock.taskservices.map { it }.map { asynctaskstep ->
-                    val aClass = AopProxyUtils.ultimateTargetClass(asynctaskstep)
-                    val annotation = AnnotationUtils.findAnnotation(aClass, TaskStepBeskrivelse::class.java)
-                    requireNotNull(annotation) { "annotasjon mangler" }
-                    annotation.taskStepType to asynctaskstep
-                }.toMap()
-
-            val asyncTaskStep = tasktyper[task.type]
-            if (asyncTaskStep != null) {
-                kjørTaskOmIkkeAnnenTaskKjører(asyncTaskStep, task)
-            } else {
-                logger.warn("Fant ikke taskstep for tasktype ${task.type} i cucumberMock. Kjører ikke tasken.")
-            }
-
-            task
-        }
-        return taskRepositoryWrapper
-    }
-}
-
 class MockedDateProvider(val mockedDate: LocalDate) : LocalDateProvider {
     override fun now(): LocalDate = this.mockedDate
-}
-
-private fun Person.tilPersonInfo(): PersonInfo =
-    PersonInfo(
-        fødselsdato = fødselsdato,
-        navn = navn,
-        kjønn = kjønn,
-        forelderBarnRelasjon = emptySet(),
-        forelderBarnRelasjonMaskert = emptySet(),
-        adressebeskyttelseGradering = null,
-        bostedsadresser = emptyList(),
-        sivilstander = emptyList(),
-        opphold = emptyList(),
-        statsborgerskap = emptyList(),
-        dødsfall = null,
-        kontaktinformasjonForDoedsbo = null,
-    )
-
-fun opprettEllerHentVedtak(
-    dataFraCucumber: BegrunnelseTeksterStepDefinition,
-    behandlingId: Long,
-): Vedtak {
-    val vedtakForBehandling =
-        dataFraCucumber.vedtaksliste.find { it.behandling.id == behandlingId }
-            ?: lagVedtak(dataFraCucumber.behandlinger[behandlingId]!!)
-
-    if (vedtakForBehandling !in dataFraCucumber.vedtaksliste) {
-        dataFraCucumber.vedtaksliste.add(vedtakForBehandling)
-    }
-
-    return vedtakForBehandling
 }
