+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/cucumber/mock/komponent_mocks/MockOppgaveService.kt
@@ -1,0 +1,11 @@
+﻿package no.nav.familie.ba.sak.cucumber.mock
+
+import io.mockk.every
+import io.mockk.mockk
+import no.nav.familie.ba.sak.integrasjoner.oppgave.OppgaveService
+
+fun mockOppgaveService(): OppgaveService {
+    val oppgaveService = mockk<OppgaveService>()
+    every { oppgaveService.opprettOppgaveForManuellBehandling(any(), any(), any(), any()) } returns ""
+    return oppgaveService
+}
