+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/cucumber/mock/komponent_mocks/MockPersonidentService.kt
@@ -1,0 +1,22 @@
+﻿package no.nav.familie.ba.sak.cucumber.mock
+
+import io.mockk.every
+import io.mockk.mockk
+import no.nav.familie.ba.sak.cucumber.BegrunnelseTeksterStepDefinition
+import no.nav.familie.ba.sak.kjerne.personident.PersonidentService
+
+fun mockPersonidentService(
+    dataFraCucumber: BegrunnelseTeksterStepDefinition,
+    småbarnstilleggBehandlingId: Long,
+): PersonidentService {
+    val personidentService = mockk<PersonidentService>()
+    every { personidentService.hentOgLagreAktør(any(), any()) } answers {
+        val personIdent = firstArg<String>()
+        dataFraCucumber.persongrunnlag[småbarnstilleggBehandlingId]!!.personer.single { it.aktør.aktivFødselsnummer() == personIdent }.aktør
+    }
+    every { personidentService.hentAktør(any()) } answers {
+        val personIdent = firstArg<String>()
+        dataFraCucumber.persongrunnlag[småbarnstilleggBehandlingId]!!.personer.single { it.aktør.aktivFødselsnummer() == personIdent }.aktør
+    }
+    return personidentService
+}
