+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/cucumber/mock/komponent_mocks/MockSettPåVentRepository.kt
@@ -1,0 +1,12 @@
+﻿package no.nav.familie.ba.sak.cucumber.mock
+
+import io.mockk.every
+import io.mockk.mockk
+import no.nav.familie.ba.sak.kjerne.behandling.settpåvent.SettPåVentRepository
+
+fun mockSettPåVentRepository(): SettPåVentRepository {
+    val settPåVentRepository = mockk<SettPåVentRepository>()
+
+    every { settPåVentRepository.findByBehandlingIdAndAktiv(any(), any()) } returns null
+    return settPåVentRepository
+}
