+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/cucumber/mock/komponent_mocks/MockSimuleringService.kt
@@ -1,0 +1,11 @@
+﻿package no.nav.familie.ba.sak.cucumber.mock
+
+import io.mockk.every
+import io.mockk.mockk
+import no.nav.familie.ba.sak.kjerne.simulering.SimuleringService
+
+fun mockSimuleringService(): SimuleringService {
+    val simuleringService = mockk<SimuleringService>()
+    every { simuleringService.oppdaterSimuleringPåBehandling(any()) } returns emptyList()
+    return simuleringService
+}
